+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.3.0 (UNRELEASED) 
 
 - Officially support Javax Servlet API 4.0 (JEE 8)
+- Ensure compatibility with JRuby 10.1
 
 Breaking compatibility changes
 - Drop support for JRuby 9.3